third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/accept-ch-stickiness/same-origin-navigation-no-accept-ch.https-expected.txt

This is a testharness.js-based test.
PASS empty-ch on navigation precondition: Test that the browser does not have client hints preferences cached
PASS empty-ch on navigation set Accept-CH to non-empty first
PASS empty-ch on navigation set Accept-CH w/o header second
FAIL empty-ch on navigation got client hints according to expectations. assert_equals: message from opened page expected "PASS" but got "FAIL"
Harness: the test ran to completion.
